Move extension functions to appropriate files
