+++ app/src/main/java/com/lambdaschool/hackathon_portal/util/ViewExtensions.kt
@@ -1,7 +1,11 @@
 package com.lambdaschool.hackathon_portal.util
 
+import android.content.Context
 import android.view.View
+import android.widget.EditText
 import android.widget.TextView
+import com.afollestad.materialdialogs.MaterialDialog
+import com.afollestad.materialdialogs.datetime.datePicker
 
 fun View.visGone() {
     this.visibility = View.GONE
@@ -17,4 +21,15 @@ fun View.visVisible() {
 
 fun TextView.text(string: String) {
     this.text = string
+}
+
+fun EditText.setClickListenerToOpenDatePickerAndSetTextToDate(context: Context) {
+    this.setOnClickListener {
+        MaterialDialog(context).show {
+            datePicker { _, datetime ->
+                this@setClickListenerToOpenDatePickerAndSetTextToDate
+                    .setText(datetime.formatCalendarToString())
+            }
+        }
+    }
 }
